requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.1
channels==2.4.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-rest-framework==0.1.0
django-rest-jwt==0.1.46
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==1.2.0
gunicorn==20.0.4
html5lib==1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.47
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.4
zope.interface==5.1.0
django-multiselectfield==0.1.12
